Reorganize the tests a bit
